miłujcie się: beautify, level 3 ..
i prefer to have simple beams in simple vocal music
to make sight-reading easier for beginners.
It would be nice to have some kind of beaming setting
for vocal music.  Or at least to be able to put just one
such override in the \layout block, instead of having to
write it in every voice, with every time signature.